Removed text offset bug
Part of the issue still exists in VOD and Audio, for some reason focus
Events are not being passed through collectionviews when the menu or
play/pause buttons are pressed so all collection views are being made
into SuperCollectionViews that have this problem customly fixed.
Converted keyPhrasesIndex reference file into english so we can easily
understand it.